+++ src/rules/comment-word-blacklist/__tests__/index.js
@@ -11,21 +11,39 @@ testRule(rule, {
   accept: [ {
     code: "/* comment */",
   }, {
-    code: "/*! bad-word */",
-  }, {
-    code: "/*# bad-word */",
-  }, {
-    code: "/** bad-word **/",
-  }, {
-    code: "/*** bad-word ***/",
+    code: "/*# bad-word */", // ignore sourcemaps
   } ],
 
-  reject: [{
+  reject: [ {
+    code: "/* Comment with bad-word  */",
+    message: messages.rejected("bad-word"),
+    line: 1,
+    column: 1,
+  },
+  {
     code: "/* bad-word */",
     message: messages.rejected("bad-word"),
     line: 1,
     column: 1,
-  }],
+  },
+  {
+    code: "/*** bad-word ***/",
+    message: messages.rejected("bad-word"),
+    line: 1,
+    column: 1,
+  },
+  {
+    code: "/*! bad-word */",
+    message: messages.rejected("bad-word"),
+    line: 1,
+    column: 1,
+  },
+  {
+    code: "/** bad-word **/",
+    message: messages.rejected("bad-word"),
+    line: 1,
+    column: 1,
+  } ],
 })
 
 testRule(rule, {
@@ -56,6 +74,8 @@ testRule(rule, {
   }, {
     code: "/*# sourcemap */",
   }, {
+    code: "/*# sourcemap bad-word */",
+  }, {
     code: "a { color: pink; /* comment */\ntop: 0; }",
   }, {
     code: "a {} /* comment */",
@@ -69,20 +89,10 @@ testRule(rule, {
     code: "/* tOdO: comment */",
   }, {
     code: "/* Todo: comment */",
-  }, {
-    code: "/* Comment with bad-word */",
-  }, {
+  },  {
     code: "/*! Todo: comment */",
   }, {
     code: "/*# Todo: comment */",
-  }, {
-    code: "/*! bad-word */",
-  }, {
-    code: "/*# bad-word */",
-  }, {
-    code: "/** bad-word **/",
-  }, {
-    code: "/*** bad-word ***/",
   }, {
     code: "/** TODO: comment **/",
   }, {
@@ -141,6 +151,26 @@ testRule(rule, {
     column: 1,
   }, {
     code: "/* bad-word */",
+    message: messages.rejected("bad-word"),
+    line: 1,
+    column: 1,
+  }, {
+    code: "/* Comment with bad-word */",
+    message: messages.rejected("bad-word"),
+    line: 1,
+    column: 1,
+  }, {
+    code: "/*! copyright bad-word */",
+    message: messages.rejected("bad-word"),
+    line: 1,
+    column: 1,
+  }, {
+    code: "/** bad-word **/",
+    message: messages.rejected("bad-word"),
+    line: 1,
+    column: 1,
+  }, {
+    code: "/*** bad-word ***/",
     message: messages.rejected("bad-word"),
     line: 1,
     column: 1,